improve mpu9255 driver noise and cpu consumption
The previous code did not make use of the fifo for the magnetometer readings.
As the magnetometer is the noisiest sensor, it needs the most samples averaged
to get lower noise.  This required me to run the kalman filter at 100hz which
used a considerable amount of cpu (30% or more on 700mhz raspberry zero)
which is a problem for a realtime priority task.

It was also a problem to ensure the i2c bus is read at 100hz every time.
Now it can be read much slower reducing bus overhead with the same noise
level as it still reads the same number of samples.

Furthermore, there was a serious bug, in that occasionally, the high
and low bytes of a magnetometer sample were from different samples giving
occasional spikes in readings if the sensor value was near zero.  This
was because of an incorrectly set register in the configuration to not
hold the sample, but the problem also disappears now that the fifo is used.

This commit, uses the mpu9255's fifo for accel, gyro and now compass data too.

The kalman filter can be run at a much lower rate now, with the same high
sample rate of the sensors.  I am sampling at 100hz but running the kalman
filter (and therefore reading the fifo contents) at 10hz on a raspberry pi
with 1% cpu usage.